+++ src/CMakeLists.txt
@@ -1,5 +1,5 @@
-add_executable(main main.c main.h database.c calculations.c CreateMenu.c CreateMenu.h ListEditor.c ListEditor.h search.c search.h ListSearch.c ListSearch.h shoppingListLoad.c shoppingListLoad.h mergeSort.c mergeSort.h calculations.h)
-add_executable(database database.c database.h)
-add_executable(search search.c search.h database.c CreateMenu.c)
+add_executable(main main.c main.h Database.c Calculations.c CreateMenu.c CreateMenu.h ListEditor.c ListEditor.h Search.c Search.h ListSearch.c ListSearch.h ShoppingListLoad.c ShoppingListLoad.h MergeSort.c MergeSort.h Calculations.h)
+add_executable(database Database.c Database.h)
+add_executable(search Search.c Search.h Database.c CreateMenu.c)
 file(COPY data DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
-add_executable(listSearch listSearch.c listSearch.h database.c shoppingListLoad.c search.c)
+add_executable(listSearch listSearch.c listSearch.h Database.c ShoppingListLoad.c Search.c)
