+++ src/Calculations.h
@@ -5,7 +5,14 @@
 #ifndef BARGAINHUNTING_CALCULATIONS_H
 #define BARGAINHUNTING_CALCULATIONS_H
 
-#include "database.h"
+#include "Database.h"
+#include "stdio.h"
+#include "Database.h"
+#include "Search.h"
+#include "CreateMenu.h"
+#include "string.h"
+#include "MergeSort.h"
+#define MAX_STR_LEN 50
 
 void Calculations (product_t **storeProducts, int numberOfStores, int numberOfSearches);
 double TotalPrice(product_t* products);
