+++ src/CreateMenu.h
@@ -4,11 +4,12 @@
 
 #ifndef BARGAINHUNTING_CREATEMENU_H
 #define BARGAINHUNTING_CREATEMENU_H
-
+#include <stdlib.h>
+#include <stdio.h>
 #define MAX_OPTIONS 20
 #define MAX_STRLEN 255
 
-#endif //BARGAINHUNTING_CREATEMENU_H
-
 int ScanInput(int items);
 int CreateMenu(char options[MAX_OPTIONS][MAX_STRLEN]);
+
+#endif //BARGAINHUNTING_CREATEMENU_H
