+++ src/ListEditor.h
@@ -4,6 +4,13 @@
 
 #ifndef BARGAINHUNTING_LISTEDITOR_H
 #define BARGAINHUNTING_LISTEDITOR_H
+#include "CreateMenu.h"
+#include "ListSearch.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <dirent.h>
+
 void ListEditor();
 
 #endif //BARGAINHUNTING_LISTEDITOR_H
