+++ src/ListSearch.h
@@ -5,11 +5,15 @@
 #ifndef BARGAINHUNTING_LISTSEARCH_H
 #define BARGAINHUNTING_LISTSEARCH_H
 
-#include "database.h"
+#include "Database.h"
 #include "ListEditor.h"
-#include <stdio.h>
+#include "Search.h"
+#include "ShoppingListLoad.h"
+#include "Calculations.h"
+#include "CreateMenu.h"
 #include <stdlib.h>
 #include <string.h>
+#include <stdio.h>
 #include <dirent.h>
 
 void ListSearch(FILE* filePtr);
