+++ src/ShoppingListLoad.h
@@ -1,0 +1,16 @@
+//
+// Created by Ida Marie Hem and Alexander Winther Moldt on 07-12-2022.
+//
+
+#ifndef BARGAINHUNTING_SHOPPINGLISTLOAD_H
+#define BARGAINHUNTING_SHOPPINGLISTLOAD_H
+#include <string.h>
+#include <stdio.h>
+#include <stdlib.h>
+
+#define SIZE_JUMPS 1
+#define MAX_BUFFER 100
+
+char **LoadFile(FILE* file);
+
+#endif //BARGAINHUNTING_SHOPPINGLISTLOAD_H
