+++ src/data/shopping_list/sinas_shopping_list.txt
@@ -1,3 +1,3 @@
-spaghetti
+noodles
 milk
 carrots
