+++ src/main.h
@@ -4,5 +4,10 @@
 
 #ifndef BARGAINHUNTING_MAIN_H
 #define BARGAINHUNTING_MAIN_H
+#include "CreateMenu.h"
+#include "Database.h"
+#include "ListEditor.h"
+#include "Search.h"
+#include <stdio.h>
 
 #endif //BARGAINHUNTING_MAIN_H
